move BaseTestCase class to test root dir
