+++ tests/Functional/Auth/LoginTest.php
@@ -3,7 +3,7 @@
 namespace Tests\Functional\Auth;
 
 use Conduit\Models\User;
-use Tests\Functional\BaseTestCase;
+use Tests\BaseTestCase;
 use Tests\UseDatabaseTrait;
 
 class LoginTest extends BaseTestCase
