+++ tests/Functional/HomepageTest.php
@@ -2,6 +2,8 @@
 
 namespace Tests\Functional;
 
+use Tests\BaseTestCase;
+
 class HomepageTest extends BaseTestCase
 {
     /**
