+++ tests/Functional/Profile/ProfileTest.php
@@ -2,7 +2,7 @@
 
 namespace Tests\Profile;
 
-use Tests\Functional\BaseTestCase;
+use Tests\BaseTestCase;
 use Tests\UseDatabaseTrait;
 
 class ProfileTest extends BaseTestCase
